Refactor and expand
