+++ built-in/built-in.h
@@ -6,7 +6,7 @@
 /*   By: achane-l <achane-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/28 14:56:36 by achane-l          #+#    #+#             */
-/*   Updated: 2022/04/02 02:01:56 by achane-l         ###   ########.fr       */
+/*   Updated: 2022/05/18 15:38:52 by achane-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,15 @@
 
 #include "../minishell.h"
 
-void	free_env(t_env *env);
 int     ft_pwd(void);
 int	    ft_env(char **env);
+//export et export utils
 int	    ft_export(t_data *data, char *str);
+char    *get_syntax(char *str);
+int     env_line_already_exist(t_data *data, t_env *new);
+void	add_env_line(t_env *env, t_env *new);
+
+void	free_env(t_env *env);
 int     ft_unset(t_data *data, char *str);
 int     ft_echo(char **args);
 int     ft_cd(t_data *data, char **params);
