+++ exec_files/exec_files.h
@@ -6,7 +6,7 @@
 /*   By: achane-l <achane-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/21 11:58:21 by achane-l          #+#    #+#             */
-/*   Updated: 2022/04/03 17:10:28 by achane-l         ###   ########.fr       */
+/*   Updated: 2022/05/17 19:08:33 by achane-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,17 +30,17 @@ typedef struct s_cmd
 int		count_arg_in_command(t_token *token);
 char	**get_cmd(t_token *token);
 // get_file.c
-int		here_doc(char *word);
-int		open_file(char *filename, int type);
-int		get_infile(t_token *token);
+int		here_doc(t_data *data, char *word);
+int		open_file(t_data *data, char *filename, int type);
+int		get_infile(t_data *data, t_token *token);
 int		get_outfile(t_token *token);
-void	add_file_to_cmd(t_cmd *cmd, t_token *token);
+void	add_file_to_cmd(t_data *data, t_cmd *cmd, t_token *token);
 
 // init_cmds.c
-t_cmd	*create_cmd(t_token *token);
+t_cmd	*create_cmd(t_data *data, t_token *token);
 void	add_cmd(t_cmd *lst_cmd, t_cmd *new_cmd);
 void	free_cmd(t_cmd **cmds);
-t_cmd	*init_cmds(t_token *token);
+t_cmd	*init_cmds(t_data *data, t_token *token);
 
 // paths
 int		check_env_val(char *src, char *find);
