+++ exec_files/init_cmds.c
@@ -6,20 +6,20 @@
 /*   By: achane-l <achane-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/01 19:13:18 by achane-l          #+#    #+#             */
-/*   Updated: 2022/03/08 14:28:35 by achane-l         ###   ########.fr       */
+/*   Updated: 2022/05/17 19:02:23 by achane-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "./exec_files.h"
 
-t_cmd	*create_cmd(t_token *token)
+t_cmd	*create_cmd(t_data *data, t_token *token)
 {
 	t_cmd	*cmd;
 
 	cmd = malloc(sizeof(t_cmd));
 	if (cmd == NULL)
 		return (NULL);
-	add_file_to_cmd(cmd, token);
+	add_file_to_cmd(data, cmd, token);
 	cmd->cmd_param = get_cmd(token);
 	cmd->next = NULL;
 	return (cmd);
@@ -32,7 +32,7 @@ void	add_cmd(t_cmd *lst_cmd, t_cmd *new_cmd)
 	lst_cmd->next = new_cmd;
 }
 
-t_cmd	*init_cmds(t_token *token)
+t_cmd	*init_cmds(t_data *data,t_token *token)
 {
 	t_cmd	*lst_cmd;
 	t_cmd	*new_cmd;
@@ -42,7 +42,7 @@ t_cmd	*init_cmds(t_token *token)
 	{
 		if (!token->prev || token->prev->type == PIPE)
 		{
-			new_cmd = create_cmd(token);
+			new_cmd = create_cmd(data, token);
 			if (!new_cmd)
 			{
 				free_cmd(&new_cmd);
