+++ main.c
@@ -6,7 +6,7 @@
 /*   By: achane-l <achane-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/04 14:23:19 by esafar            #+#    #+#             */
-/*   Updated: 2022/04/29 00:26:20 by achane-l         ###   ########.fr       */
+/*   Updated: 2022/05/18 14:57:37 by achane-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,9 +41,9 @@ int main(int ac, char **av, char **env)
 		if ((ft_strlen(str) != 0) && parse(&data, str))
 		{
 			//print_token_list(data.begin);
-			data.cmd = init_cmds(data.begin);
+			data.cmd = init_cmds(&data, data.begin);
 			exec(&data);
-			free_lst(&data);
+			free_token_lst(&data);
 		}
 		free(str);
 	}
